Added boss "Fenrique"
- Added boss "Fenrique", featuring a tank hero and 4 dps heroes.
- Added boss selection menu.
- Modified game and boss initialization to include the boss to fight as
an argument.
- Added threat management options for bossfights.
- getRandomHero now takes a list of target roles as an argument. If this
is not null, a hero will be chosen based on the passed roles, taking
priority for the order of roles listed. For example, if you pass an
array where the first element consists of the tank role and the second
consists of the dps role, any available tanks will be chosen before a
dps will be chosen.
- Various style tweaks.